+++ package.json
@@ -23,5 +23,8 @@
   "engines": {
     "node": "18",
     "npm": "9"
-  }
+  },
+
+  "type": "module"
+
 }
